2025-2-18-2

+++ w-monorepo/apps/management-price/src/app/pages/platform/platform.component.ts
@@ -23,7 +23,7 @@ export class PlatformComponent implements OnInit {
   protected platforms: IPlatform[] = [];
   protected totalPages = 0;
   protected currentPage = 0;
-  protected pageSize = 2; // 每页显示数量
+  protected pageSize = 15; // 每页显示数量
   public isAddModalVisible = false; // 控制新增商品表单显示
   public addPlatform: IAddPlatform = { name: '' }; // 用于绑定新增商品数据
   public isEditModalVisible = false; // 控制编辑商品表单显示
@@ -116,15 +116,23 @@ export class PlatformComponent implements OnInit {
 
   // 删除商品
   public deletePlatform(id: string): void {
-    this.platformHttpService.deletePlatform(id).subscribe(
-      () => {
-        this.platforms = this.platforms.filter(
-          (platform) => platform.id !== id
-        ); // 从列表中删除商品
-      },
-      (error) => {
-        console.error('删除商品失败', error);
-      }
-    );
+    // 弹出二次确认框
+    const isConfirmed = window.confirm('确定要删除该平台吗？ id 为：' + id);
+
+    if (isConfirmed) {
+      this.platformHttpService.deletePlatform(id).subscribe(
+        () => {
+          this.platforms = this.platforms.filter(
+            (platform) => platform.id !== id
+          ); // 从列表中删除商品
+        },
+        (error) => {
+          console.error('删除平台失败', error);
+        }
+      );
+    } else {
+      // 用户点击取消，不做任何操作
+      console.log('删除操作已取消');
+    }
   }
 }
